add movie seperated
