+++ package.json
@@ -23,6 +23,8 @@
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.9",
     "@vitejs/plugin-react": "^3.0.0",
+    "eslint": "^8.0.0",
+    "eslint-config-react-app": "^7.0.1",
     "vite": "^4.0.0"
   }
 }
