+++ src/.eslintrc.json
@@ -1,0 +1,4 @@
+{
+    "extends": "react-app"
+  }
+
