+++ src/MovieList.jsx
@@ -1,34 +1,10 @@
-import { useState } from "react";
-import TextField from '@mui/material/TextField';
-import Button from '@mui/material/Button';
 import { Movie } from "./Movie";
+import { AddMovie } from "./AddMovie";
 
 export function MovieList({movieList, setmovieList}) {
- 
-  const [Name, setName] = useState("");
-  const [Poster, setPoster] = useState("");
-  const [Rating, setRating] = useState("");
-  const [Summary, setSummary] = useState("");
-
   return (
     <div>
-      <div className="add-movie-form">
-        <TextField onChange={(event) => setName(event.target.value)} label="Name" variant="outlined" />
-        <TextField onChange={(event) => setRating(event.target.value)} label="Rating" variant="outlined" />
-        <TextField onChange={(event) => setPoster(event.target.value)} label="Poster" variant="outlined" />
-        <TextField onChange={(event) => setSummary(event.target.value)} label="Summary" variant="outlined" />
-
-        <Button variant="contained" onClick={() => {
-          const newMovie = {
-            name: Name,
-            poster: Poster,
-            rating: Rating,
-            summary: Summary
-          };
-          setmovieList([...movieList, newMovie]);
-        }}>Add Movie</Button>
-        {/* <button >Add Movie</button> */}
-      </div>
+      {/* <AddMovie movieList={movieList} setmovieList={setmovieList} /> */}
       <div className="movie-list">
         {movieList.map((mv, index) => (
           <Movie key={index} movie={mv} id={index} />
@@ -37,3 +13,4 @@ export function MovieList({movieList, setmovieList}) {
     </div>
   );
 }
+
